requirement.txt
python-telegram-bot==20.4
aiofiles
python-dotenv
